Update react router
