--- space-travel/src/App.js
@@ -1,10 +1,0 @@
-import Nav from './component/Nav';
-import './App.css';
-
-function App() {
-  return (
-    <Nav />
-  );
-}
-
-export default App;
