--- space-travel/src/component/Nav.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import logo from './planet.png'
-
-const Nav = () => (
-  <nav>
-    <img src={logo} alt="space-logo" />
-    <h3>Space Traveler&apos Hub</h3>
-  </nav>
-);
-
-export default Nav;
