+++ src/component/Missions.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+const Missions = () => (
+  <h1>Missions page</h1>
+);
+export default Missions;
