+++ src/component/MyProfile.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+const MyProfile = () => (
+  <h1>My profile page</h1>
+);
+export default MyProfile;
