+++ src/component/Rockets.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+const Rockets = () => (
+  <h1>Rockets page</h1>
+);
+export default Rockets;
